+++ .gitignore
@@ -3,4 +3,4 @@
 models
 __pycache*
 data
-
+.ipynb_checkpoints/
